Add the table to see assignments for a class and update the SQL and Springboot to handle the grade being passed as a double instead of an int

+++ SQL Server/CREATE.sql
@@ -95,7 +95,7 @@ CREATE TABLE student_assignment (
 	id INT NOT NULL AUTO_INCREMENT, 
     student_course_id INT NOT NULL, 
     course_assignment_id INT NOT NULL, 
-    grade INT, 
+    grade DOUBLE, 
     feedback VARCHAR(225), 
     PRIMARY KEY (id)
 );
